Refactor out no longer abstract base.
